Extend prototypes in a more encapsulated way.
