+++ tools/Unicode/CaseInsensitive/protos.d.ts
@@ -1,0 +1,9 @@
+interface String {
+    zeroPadFourDigits(): string;
+    toCodepoint(): number;
+}
+
+interface Number {
+    toUnicodeHexString(): string;
+    toCppUnicodeHexString(): string;
+}
